+++ src/test/java/lab5/players/CirceTesting.java
@@ -1,9 +1,8 @@
-package lab5;
+package lab5.players;
 
 import lab5.game.Board;
 import lab5.game.Col;
 import lab5.game.Row;
-import lab5.players.Circe;
 import lab5.game.Position;
 import org.junit.jupiter.api.Test;
 import static lab5.game.PlayerToken.O;
